refactor: split dyadtensor ad tests and simplify linalg ad macros
